refactor: move fjs init boilerplate out, shift db init
